# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Financial Constraint Analysis Platform API
  description: Enterprise-grade financial analysis platform with AI-powered insights, constraint management, and comprehensive compliance features.
  version: 1.0.0
  contact:
    name: Financial Analyzer Team
    email: support@financial-analyzer.com
  license:
    name: Proprietary
    url: https://financial-analyzer.com/license
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.financial-analyzer.com/api
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /documents:
    get:
      summary: List user documents
      tags: [Documents]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: companyId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Documents retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Document'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: Upload new document
      tags: [Documents]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, companyId, documentType]
              properties:
                file:
                  type: string
                  format: binary
                companyId:
                  type: string
                documentType:
                  type: string
                  enum: [QUARTERLY_REPORT, ANNUAL_REPORT, FINANCIAL_STATEMENT, OTHER]
      responses:
        '201':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  document:
                    $ref: '#/components/schemas/Document'
  /documents/{documentId}:
    get:
      summary: Get document details
      tags: [Documents]
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Document details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete document
      tags: [Documents]
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Document deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'

  # Constraint Management
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
        fileName:
          type: string
        fileType:
          type: string
          enum: [PDF, EXCEL, CSV]
        fileSize:
          type: integer
        documentType:
          type: string
          enum: [QUARTERLY_REPORT, ANNUAL_REPORT, FINANCIAL_STATEMENT, OTHER]
        uploadedAt:
          type: string
          format: date-time
        processedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [UPLOADED, PROCESSING, PROCESSED, FAILED]
        companyId:
          type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...